pyproject.toml
[project]
name = "wind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azula>=0.8.0",
    "cartopy>=0.23.0",
    "dask>=2026.1.2",
    "diffusers>=0.36.0",
    "einops>=0.8.2",
    "flow-matching>=1.0.10",
    "h5py>=3.14.0",
    "heavyball>=2.2.3",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
    "ipython>=8.18.1",
    "lightning>=2.6.0",
    "numcodecs>=0.12.1",
    "plotly>=6.5.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "rich>=14.3.2",
    "rootutils>=1.0.7",
    "timm>=1.0.24",
    "torch>=2.8.0",
    "torchfsm>=0.0.6",
    "torchmetrics>=1.8.2",
    "torchvision>=0.23.0",
    "wandb>=0.24.2",
    "xarray>=2024.7.0",
    "zarr>=3.1.5",
]
